Refactoring the input manager, not entirely necessary but simplifies some things.  More computer interactioin and A new LabCamera script.
