+++ Assets/GameManager/LabInputManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 52687cdc5b8c4ac42a6f6e461aa6fe84
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
