+++ Assets/GameManager/LabInputs.inputactions.meta
@@ -1,0 +1,14 @@
+fileFormatVersion: 2
+guid: 09508f536bd9efc49b14796cfe1ea544
+ScriptedImporter:
+  internalIDToNameTable: []
+  externalObjects: {}
+  serializedVersion: 2
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: 
+  script: {fileID: 11500000, guid: 8404be70184654265930450def6a9037, type: 3}
+  generateWrapperCode: 0
+  wrapperCodePath: 
+  wrapperClassName: 
+  wrapperCodeNamespace:
