+++ Assets/GameManager/TerrariumInputManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f20fcc5539594ab41937e51bd616eec9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
