+++ Assets/Scenes/Lab.unity
@@ -335,9 +335,9 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: f731f99f06301cf4199195a52fd4bade, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  cameraLoacation: {x: 18, y: 3.35, z: 10}
+  cameraLocation: {x: 0, y: 0, z: 0}
   cameraRotation: {x: 0, y: 0, z: 0, w: 1}
-  cameraSpeed: 50
+  cameraSpeed: 5
 --- !u!1 &259092770
 GameObject:
   m_ObjectHideFlags: 0
@@ -603,6 +603,23 @@ PrefabInstance:
     m_AddedGameObjects: []
     m_AddedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 4c2253518f01b3943a52397ff7079b09, type: 3}
+--- !u!1 &538424193 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 7505620091327887725, guid: a3e904a445d467d4591aebbe7cabaad3, type: 3}
+  m_PrefabInstance: {fileID: 4401467658436672895}
+  m_PrefabAsset: {fileID: 0}
+--- !u!114 &538424198
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 538424193}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 1e030fb1dd5f0b8428048d137796c6c8, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1001 &697618721
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -1394,5 +1411,8 @@ PrefabInstance:
     m_RemovedComponents: []
     m_RemovedGameObjects: []
     m_AddedGameObjects: []
-    m_AddedComponents: []
+    m_AddedComponents:
+    - targetCorrespondingSourceObject: {fileID: 7505620091327887725, guid: a3e904a445d467d4591aebbe7cabaad3, type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 538424198}
   m_SourcePrefab: {fileID: 100100000, guid: a3e904a445d467d4591aebbe7cabaad3, type: 3}
